refactor(ipc): isolate account and diagnostics validation
